src/main/webapp/app/entities/enfant/route/enfant-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IEnfant, Enfant } from '../enfant.model';
import { EnfantService } from '../service/enfant.service';

import { EnfantRoutingResolveService } from './enfant-routing-resolve.service';

describe('Enfant routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: EnfantRoutingResolveService;
  let service: EnfantService;
  let resultEnfant: IEnfant | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(EnfantRoutingResolveService);
    service = TestBed.inject(EnfantService);
    resultEnfant = undefined;
  });

  describe('resolve', () => {
    it('should return IEnfant returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultEnfant = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultEnfant).toEqual({ id: 123 });
    });

    it('should return new IEnfant if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultEnfant = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultEnfant).toEqual(new Enfant());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Enfant })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultEnfant = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultEnfant).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
